fix(db): ensure BigInteger primary keys and add pre-deploy validation script

+++ backend/app/models/ad.py
@@ -1,12 +1,12 @@
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, Boolean, DateTime
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, BigInteger
 
 from app.database import Base
 
 class AdPost(Base):
     __tablename__ = "ad_posts"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     title = Column(String(255), nullable=False)
     image_url = Column(String(500), nullable=True)
     link_url = Column(String(500), nullable=True)
